Reactify achievement detail page
This page was (and still is to some extent) very lumbering and massive. I
haven't actually stopped loading all the achievement data here but at least we
don't generate a massive HTML page any more for something like j_meka's
all seasons achievement detail page.

/api/decks now accepts and achievementKey param and will only get decks that
relate to that achievement for the specified person in the specified season(s).

In order to form the right WHERE clause for this I need to query _achievements.
Putting that in query resulted in a circular import. To untangle that I
separated query (snippets of pure text SQL) and clauses (funcs that form WHERE,
ORDER BY and LIMIT clauses, possibly by doing queries of their own) into two
separate modules.
