+++ decksite/templates/livedecktable.mustache
@@ -1,5 +1,6 @@
 <div
     class="decktable"
+    data-achievement-key="{{achievement_key}}"
     data-archetype-id="{{archetype.id}}"
     data-active-runs-text="{{^tournament_only}}{{active_runs_text}}{{/tournament_only}}"
     data-card-name="{{card.name}}"
